refactored packages to lower case
