--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Config/IConfig.java
@@ -1,22 +1,0 @@
-package org.firstinspires.ftc.teamcode.Config;
-
-public interface IConfig
-{
-    int getInt(String key);
-
-    float getFloat(String key);
-
-    double getDouble(String key);
-
-    double getChar(String key);
-
-    byte getByte(String key);
-
-    long getLong(String key);
-
-    short getShort(String key);
-
-    boolean getBoolean(String key);
-
-    String getString(String key);
-}
