+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/components/configs/ConfigParser.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Config;
+package org.firstinspires.ftc.teamcode.components.configs;
 
 import android.os.Environment;
