+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/controller/Handler.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Hardware.controller;
+package org.firstinspires.ftc.teamcode.hardware.controller;
 
 public interface Handler
 {
