+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/controller/TriggerType.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Hardware.controller;
+package org.firstinspires.ftc.teamcode.hardware.controller;
 
 /**
  * Created by EvanCoulson on 1/5/18.
