+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/autonomous/BaseAutonomousOpMode.java
@@ -1,9 +1,7 @@
-package org.firstinspires.ftc.teamcode.opmodes.Autonomous;
+package org.firstinspires.ftc.teamcode.opmodes.autonomous;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.Config.ConfigParser;
-import org.firstinspires.ftc.teamcode.systems.BaseSystems.EyeSystem;
 import org.firstinspires.ftc.teamcode.systems.IMUSystem;
 import org.firstinspires.ftc.teamcode.systems.MecanumDriveSystem;
