+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/teleop/TeleOpMode.java
@@ -1,19 +1,13 @@
 package org.firstinspires.ftc.teamcode.opmodes.TeleOp;
 
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.hardware.DcMotor;
 
-import org.firstinspires.ftc.teamcode.Config.ConfigParser;
 import org.firstinspires.ftc.teamcode.Hardware.controller.Button;
 import org.firstinspires.ftc.teamcode.Hardware.controller.Controller;
 import org.firstinspires.ftc.teamcode.Hardware.controller.Handler;
-import org.firstinspires.ftc.teamcode.opmodes.debuggers.LinearOpModeDebugger;
 import org.firstinspires.ftc.teamcode.opmodes.debuggers.TeleOpModeDebugger;
 import org.firstinspires.ftc.teamcode.systems.ArmState;
 import org.firstinspires.ftc.teamcode.systems.ArmSystem;
-import org.firstinspires.ftc.teamcode.systems.IMUSystem;
 import org.firstinspires.ftc.teamcode.systems.MecanumDriveSystem;
 
 /**
