+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/systems/base/LinearEncoderSystem.java
@@ -1,10 +1,10 @@
-package org.firstinspires.ftc.teamcode.systems.BaseSystems;
+package org.firstinspires.ftc.teamcode.systems.base;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 
-import org.firstinspires.ftc.teamcode.components.Motors.DriveMotor;
+import org.firstinspires.ftc.teamcode.components.motors.DriveMotor;
 import org.firstinspires.ftc.teamcode.components.scale.ExponentialRamp;
 import org.firstinspires.ftc.teamcode.components.scale.Point;
 import org.firstinspires.ftc.teamcode.components.scale.Ramp;
